use vex1.5 if_def P-cal base freq and sub-MHz eg 0.5 MHz PCal (Wettzell shifted PCal), at least during vex2difx stage (and verified with test case that tone selection does not change vs code prior to this commit)
git-svn-id: https://svn.atnf.csiro.au/difx@8009 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/vex2difx/trunk/vexdatamodel/vex_data.cpp
@@ -715,7 +715,7 @@ void VexData::swapPolarization(const std::string &antName)
 	}
 }
 
-void VexData::setPhaseCalInterval(const std::string &antName, int phaseCalIntervalMHz)
+void VexData::setPhaseCalInterval(const std::string &antName, float phaseCalIntervalMHz)
 {
 	for(std::vector<VexMode>::iterator it = modes.begin(); it != modes.end(); ++it)
 	{
